Mono/Linux bug fixes; enhanced realiability of gibbs reactor solver (added an option to initialize from previous solution); fixed Sensitivity Analysis utility not displaying variables (CPUI)

+++ DWSIM/Forms/Main Window/FormMain.vb
@@ -2010,20 +2010,22 @@ Public Class FormMain
             form.FormProps.DockPanel = Nothing
             form.FormCharts.DockPanel = Nothing
 
-            If Not My.Computer.Keyboard.ShiftKeyDown Then
-                If savedfromclui Then
-                    Dim myfile As String = My.Computer.FileSystem.GetTempFileName()
-                    Try
-                        Dim pnl As String = xdoc.Element("DWSIM_Simulation_Data").Element("PanelLayout").Value
-                        File.WriteAllText(myfile, pnl)
-                        form.dckPanel.LoadFromXml(myfile, New DeserializeDockContent(AddressOf Me.ReturnForm))
-                    Catch ex As Exception
-                    Finally
-                        File.Delete(myfile)
-                    End Try
-                Else
-                    Dim myfile As String = IO.Path.Combine(My.Application.Info.DirectoryPath, "layout.xml")
-                    form.dckPanel.LoadFromXml(myfile, New DeserializeDockContent(AddressOf ReturnForm))
+            If Not DWSIM.App.IsRunningOnMono Then
+                If Not My.Computer.Keyboard.ShiftKeyDown Then
+                    If savedfromclui Then
+                        Dim myfile As String = My.Computer.FileSystem.GetTempFileName()
+                        Try
+                            Dim pnl As String = xdoc.Element("DWSIM_Simulation_Data").Element("PanelLayout").Value
+                            File.WriteAllText(myfile, pnl)
+                            form.dckPanel.LoadFromXml(myfile, New DeserializeDockContent(AddressOf Me.ReturnForm))
+                        Catch ex As Exception
+                        Finally
+                            File.Delete(myfile)
+                        End Try
+                    Else
+                        Dim myfile As String = IO.Path.Combine(My.Application.Info.DirectoryPath, "layout.xml")
+                        form.dckPanel.LoadFromXml(myfile, New DeserializeDockContent(AddressOf ReturnForm))
+                    End If
                 End If
             End If
